Products branch

+++ src/main/java/com/teamtiger/productservice/products/models/ProductDTO.java
@@ -12,6 +12,6 @@ public record ProductDTO(
     double retailPrice,
 
     @NotNull
-    int weight
+    double weight
 ) {
 }
